fuzz: minimized h1_fuzz code overlap
